requirements.txt
Django==2.0.2
django-autocomplete-light==3.2.10
django-crispy-forms==1.7.2
django-extensions==2.0.7
django-localflavor==2.0
django-municipios==0.8.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.6
django-static-jquery==2.1.4
Pillow==5.0.0
pkg-resources==0.0.0
pygraphviz==1.3.1
pytz==2018.3
six==1.11.0
